style: infrastructure client > clients 패키지 경로 변경
