+++ delivery-service/src/main/java/on/logistics/deliveryservice/application/service/DeliveryRecordServiceImpl.java
@@ -15,9 +15,9 @@ import on.logistics.deliveryservice.domain.repository.DeliveryRecordRepository;
 import on.logistics.deliveryservice.exception.DeliveryRecordException;
 import on.logistics.deliveryservice.exception.DeliveryRecordExceptionCode;
 import on.logistics.deliveryservice.global.application.dtos.PageDto;
-import on.logistics.deliveryservice.infrastructure.client.hub.HubServiceClient;
-import on.logistics.deliveryservice.infrastructure.client.map.MapServiceClient;
-import on.logistics.deliveryservice.infrastructure.client.map.feign.dtos.GetEstimateInfo;
+import on.logistics.deliveryservice.infrastructure.clients.hub.HubServiceClient;
+import on.logistics.deliveryservice.infrastructure.clients.map.MapServiceClient;
+import on.logistics.deliveryservice.infrastructure.clients.map.feign.dtos.GetEstimateInfo;
 import on.logistics.deliveryservice.presentation.dtos.response.CreateDeliveryRecordResponse;
 import on.logistics.deliveryservice.presentation.dtos.response.GetDeliveryRecordResponse;
 import on.logistics.deliveryservice.presentation.dtos.response.SearchDeliveryRecordResponse;
