+++ delivery-service/src/main/java/on/logistics/deliveryservice/application/service/DeliveryServiceImpl.java
@@ -15,9 +15,9 @@ import on.logistics.deliveryservice.domain.repository.DeliveryRepository;
 import on.logistics.deliveryservice.exception.DeliveryException;
 import on.logistics.deliveryservice.exception.DeliveryExceptionCode;
 import on.logistics.deliveryservice.global.application.dtos.PageDto;
-import on.logistics.deliveryservice.infrastructure.client.hub.HubServiceClient;
-import on.logistics.deliveryservice.infrastructure.client.map.MapServiceClient;
-import on.logistics.deliveryservice.infrastructure.client.map.feign.dtos.GetDestinationInfo;
+import on.logistics.deliveryservice.infrastructure.clients.hub.HubServiceClient;
+import on.logistics.deliveryservice.infrastructure.clients.map.MapServiceClient;
+import on.logistics.deliveryservice.infrastructure.clients.map.feign.dtos.GetDestinationInfo;
 import on.logistics.deliveryservice.presentation.dtos.response.CreateDeliveryResponse;
 import on.logistics.deliveryservice.presentation.dtos.response.GetDeliveryResponse;
 import on.logistics.deliveryservice.presentation.dtos.response.SearchDeliveryResponse;
