--- delivery-service/src/main/java/on/logistics/deliveryservice/infrastructure/client/hub/feign/dtos/HubType.java
@@ -1,5 +1,0 @@
-package on.logistics.deliveryservice.infrastructure.client.hub.feign.dtos;
-
-public enum HubType {
-    HUB, SPOKE
-}
