+++ delivery-service/src/main/java/on/logistics/deliveryservice/infrastructure/clients/hub/HubServiceClient.java
@@ -1,0 +1,18 @@
+package on.logistics.deliveryservice.infrastructure.clients.hub;
+
+import java.util.UUID;
+import on.logistics.deliveryservice.infrastructure.clients.hub.feign.dtos.GetHubInfo;
+import on.logistics.deliveryservice.infrastructure.clients.hub.feign.dtos.GetMiddleHubPageInfo;
+import on.logistics.deliveryservice.infrastructure.clients.hub.feign.dtos.GetSpokeHubInfo;
+import on.logistics.deliveryservice.infrastructure.clients.hub.feign.dtos.HubType;
+import org.springframework.web.bind.annotation.RequestParam;
+
+public interface HubServiceClient {
+
+    GetHubInfo getHubInfo(UUID hubId);
+
+    GetMiddleHubPageInfo searchHubs(@RequestParam("type") HubType type);
+
+    GetSpokeHubInfo getSpokeHubInfo(UUID hubId);
+
+}
