+++ delivery-service/src/main/java/on/logistics/deliveryservice/infrastructure/clients/hub/feign/dtos/HubInfo.java
@@ -1,4 +1,4 @@
-package on.logistics.deliveryservice.infrastructure.client.hub.feign.dtos;
+package on.logistics.deliveryservice.infrastructure.clients.hub.feign.dtos;
 
 public record HubInfo(String id, String name, String type, String address, String latitude,
                       String longitude) {
