+++ delivery-service/src/main/java/on/logistics/deliveryservice/infrastructure/clients/hub/feign/dtos/HubType.java
@@ -1,0 +1,5 @@
+package on.logistics.deliveryservice.infrastructure.clients.hub.feign.dtos;
+
+public enum HubType {
+    HUB, SPOKE
+}
