+++ delivery-service/src/main/java/on/logistics/deliveryservice/infrastructure/clients/map/feign/dtos/GetDestinationInfo.java
@@ -1,4 +1,4 @@
-package on.logistics.deliveryservice.infrastructure.client.map.feign.dtos;
+package on.logistics.deliveryservice.infrastructure.clients.map.feign.dtos;
 
 public record GetDestinationInfo(String roadAddress, String jibunAddress, String longitude,
                                  String latitude) {
